Create README.md
The app is intended to help people to organise their trips beforehand.
On the main menu, you can select your destination among the 3 possible options and the date of your trip. If the date is not valid, a popup window will show up with a warning.

Pressing the button "Update Trip", the text view "Your trip to (destination)!" changes and you can see, at the bottom, how many days your trip is.

The button also changes the location of the pin on the map shown in the middle of the screen and if you click on the pin, it takes you to Google Maps app, where you can check the nearest restaurants and attractions in the chosen city. If your trip is in 7 days or less, you get a notification as a reminder with the remaining days to your trip.

When you click on the checklist image, it takes you to another screen where you can select the type of trip you intend to make (beach, business, mountain or cultural) and depending on your destination the app will show you only the possible options for that location (for example, if you are going to London it will not show you options for beach or mountain). After selecting it and pressing the "See my checklist" button, you are provided with a list of important things to pack for your trip. Your selections will be saved unless you press the button again.

In the tips section, you can see tips for the destination you have previously selected in the menu. You are also able to choose a different location, in case of curiosity, by selecting a new destination in the spinner and pressing the "Show me new tips!" button. You should scroll down to see all the tips.

The app can be used in both portrait and landscape modes.